# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  Day: string;
  Time: string;
  Title: string;
  Description: string;
  Tag: string;
}

export interface WhatsOnResponse {
  events: Event[];
}

export interface MenuItem {
  name: string;
  desc: string;
  price: string;
}

export interface MenuSection {
  [key: string]: MenuItem[];
}

export interface MenuResponse {
  sections: MenuSection;
}

export type GetWhatsOnParams = {
  /**
   * Google Spreadsheet ID
   */
  sheetId: string;
};

export type GetMenuParams = {
  /**
   * Google Spreadsheet ID
   */
  sheetId: string;
};