# lib/api-zod/src/generated/types/menuSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItem } from "./menuItem";

export interface MenuSection {
  [key: string]: MenuItem[];
}